requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0
pydantic==2.5.0
pinecone-client==3.0.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
protobuf>=4.21.0

